Phase3_Voice/requirements.txt
livekit-agents
livekit-plugins-deepgram
livekit-plugins-groq
livekit-plugins-cartesia
livekit-plugins-silero
python-dotenv>=1.0.0
